pyproject.toml
[tool.poetry]
name = "data-science-setup"
version = "0.1.0"
description = ""
authors = ["Roger Galindo <rogherschrodinger@gmail.com>"]
readme = "README.md"
packages = [{include = "data_science_setup"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.3"
scikit-learn = "^1.2.1"
pre-commit = "^3.0.4"
dvc = "^2.45.1"
hydra-core = "^1.3.1"
matplotlib = "^3.7.0"
seaborn = "^0.12.2"
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pdoc3 = "^0.10.0"
pca = "^1.9.1"
great-expectations = "^0.15.50"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
